# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plumbing Business AI Platform API
  description: |
    Comprehensive API for the production-ready plumbing business AI platform.
    
    This API provides endpoints for managing customers, jobs, conversations, AI interactions,
    emergency responses, scheduling, pricing, warranties, and system monitoring.
    
    ## Features
    - **Customer Management**: Complete customer lifecycle management
    - **AI-Powered Conversations**: Intelligent message processing and response generation
    - **Emergency Response**: Real-time emergency detection and routing
    - **Dynamic Pricing**: Intelligent pricing based on multiple factors
    - **Scheduling Optimization**: Advanced technician scheduling and route optimization
    - **Warranty Management**: Comprehensive warranty tracking and claims processing
    - **Performance Monitoring**: Real-time system health and performance metrics
    
    ## Authentication
    All endpoints require JWT-based authentication unless otherwise specified.
    Include the token in the Authorization header: `Bearer {token}`
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Standard endpoints: 1000 requests per hour per user
    - Search endpoints: 100 requests per minute per user
    - Webhook endpoints: 10,000 requests per hour
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": {
...
        }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user credentials and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: admin@plumbingcompany.com
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: string
                    example: 24h
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/register:
    post:
      tags: [Authentication]
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: User registered successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidate current session token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logged out successfully

  # Customer Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, dispatcher, technician, customer_service]
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    UserRegistration:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        name:
          type: string
          minLength: 2
        role:
          type: string
          enum: [dispatcher, technician, customer_service]
        permissions:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error - Invalid input format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...